feat: renamed config.ini to config.sample.ini and fixed error when using old config file
The newer config file is being renamed to config.sample.ini in order to avoid overriding existing configs. An update has also been made so that the old config is now compatible with the new code. If a config file does not exist on running LinearStitch, it will be created from the sample file. Documenation has been added to the README.